Fix a linking cache failure
This fixes the bug where changing the pconfigure arguments such that
only a "cp" was required to change the resulting binary caused the "cp"
to never actually get run.  This happened because the only thing that
changed was the Makefile and that was no longer being depended upon.

The fix is to just depend on the Makefile for all the "cp" operations.