Add a --verbose flag to pconfigure
Passing "--verbose" to pconfigure means that it will print out the
exact commands used to build everything instead of printing out a pretty
compilation message.  The main goal here is to support people who want
to actually see the compiler invocations scroll across the screen.  This
is mostly useful for distributions that want to more easily debug build
failures.

There were some changes to the makefile generator to enable this, but
it required to changes to the language support files.  I also made the
verbose mode pretty up the output of the language modules by stripping
backslashes at the end of lines (and joining them back together).