Add a --config option to pconfigure
This option allows one to specify an extra (or multiple extra)
Configfiles that will get added to the build (in addition to all the
regular files, Configfiles{/local,}).  These files will get added after
Configfiles/local but before Configfile.

The general idea is that different configurations (enabling debug, for
example) can be comitted to the repository to make it easier for users
to learn about the build system.

As there are probably many configfiles now, I moved them to their own
directory, Configfiles.