Add some library support
It's now possible (in C, at least) to build libraries.  As of now
there is no support for resolving binary to library dependencies in
the same project (ie, writing a program that depends upon its own
library) as that requires a bit more work.

Additionally, this doesn't set -rpath for C executables as I need to
implement a language-specific install phase (to then remove the rpath)
where's right now it's just a "cp".