Add DEPLIBS
This is a way to specify that a binary depends on a library that is
built as part of this project.  Right now it only works for C code,
can't be set globally, doesn't do any error checking, and won't work
if the library output directory is changed.  All these bugs are fixable
(some easier than others) but I'm too lazy to fix them for now.

This also adds -rpath to all C executables (which is a big part of why
only C supports it -- though I suppose C++ does too, as they're the
same here).  There isn't any support for removing this -rpath, so these
executables shouldn't be "make install"ed until I fix that bug.