Add DEPLIBS Makefile dependency
It turns out I forgot to actually add a dependency on this, I had just
added the link step.  Whoops!  The bug is fixed now, I just added the
library as a dependency to every binary that links it.  This is the
proper solution, and is the reason I decided to require explicit forms
for inter-project library dependencies -- I guess I just forgot about
it.