Add proper "make install" support back
This now links everything twice: once if we're trying to run it inside
the current directory, and once again when we're trying to install the
files into the system.  Right now the only difference is that the local
target adds a "-rpath" and the system one doesn't.  This support only
exists for C (and derivitives) and only exists for executables (ie, not
libraries).

This required another bit of code duplication, which I don't like all
that much.  I think I'm going to clean everything up a bit before
trying to commit this.  This stuff also requires a bit of testing :).