Fix an install-related bug
I forgot that we still need to set the link-time library path for the
install binaries, just not the run-time library path -- clearly just an
oversight because they were done one the same line.

Though this does hint that I need a test suite for pconfigure.  I've
been considering adding a simple test framework to pconfigure for a
while now, and I think it's time.  Though I think I'll release 1.6.1
before doing this.