Properly support linker scripts
It turns out this was actually super easy, but for some reason I was
being a lazy bum and not supporting them in a clean manner.  That's
all fixed now, which means that linker scripts are handled by the
dependency system.

Note that there's an explicit assumption here (at least, implicit in
the code) that linker scripts don't have any dependencies of their
own.  I'm not sure if this is true or not, but for right now I'm just
going to go ahead and assume that nobody in their right minds would
ever make a linker script that complicated.  I'm sure I'll be proven
wrong quite soon, though :).