Add pgcc-config
The idea behind this is that sometimes you want to designate a GCC
minor version, but you don't really care about the patch version.
This lets you select a GCC that works for you.

Now that I think about it, I should probably extend this a bit, but I
guess it's OK for now...