Add scala helper scripts
There's two of these helper scripts: one that's designed to emulate a
compiler and one that's designed to emulate a linker.  The compiler
takes a Scala file and compiles it into a jar archive, while the
linker merges a number of jar archives together.

I'd like to make the linker script output a bash archive so the output
is actually executable, but I'm on a plane and I don't remember
exactly how these bash archives are generated.

Also, there's no support in either of these for commandline argument
parsing.  The linker should probably support some sort of library
specification syntax (I'm hoping to make it look just like GCC), but I
have to think about that for a few minutes.