Add Scala support
This supports building Scala projects.  Note that it doesn't handle
dependencies or incremental compilation because, as far as I can tell,
Scala doesn't support these.  It would be great if somebody could
somehow figure out how do to this, but I guess it's just not that high
on the list of priorities for the Scala people.

Note that this means that there won't be any parallel Scala
compilation, which is a huge pain.