Call "scala" "scala", not "c"
This was messing with the code that detects when a language is added
twice.  When I added Scala before C, the loop to detect that a
language was being added twice would look at Scala, see that it's
called "c", and fail.

The fix is quite simple: just call Scala "scala"!