Make talloc dependency optional
The idea is that some people might not actually have talloc.  In order
to avoid forcing them to depend on talloc I'm instead simply include a
version of talloc that I staticly link into pconfigure in this case.

There is still a dependency on clang here.  I think the best way to
avoid the clang dependency is to write a mini-clang (ie, call GCC's
dependency generator a few times) that I use when clang isn't
availiable.