Add ptest, a testing framework
This is designed to be the simplest testing framework possible: it
simply runs a list of tests, reporting if they pass or fail.  The idea
is that I don't want to impose any sort of testing framework on the
user but instead just want a nice clean interface for running tests.

The nice part is that tests are just binaries, so they can by built by
pconfigure along with the rest of your binaries.  This is currently a
double-edged sword: it's pretty but it's really showing the
deficiencies in my BASH build support.  I guess that's a good excuse
to fix up my BASH support... :).