Change ptest reporting method
This changes ptest to be more in line with the rest of pconfigure:
"make check" will run the tests, while "ptest" will report the test
results.  The idea is that I don't want Makefiles to just become
scripts, which is what would happen if you have "make check" report
the results from the tests.