Support #include in pbashc
This works just like #include in C, the current line is replaced with
the requested file.  This allows shared code between bash scripts
without having to deal with any sort of nasty dependencies (ie,
discovering system paths).  This, of course, comes with the drawback
that the files balloon in size, but that's not really a big deal as
they're just bash scripts.

Note that pconfigure currently doesn't detect dependencies for these