Add libpinclude and pinclude
libpinclude is a simple library for listing the #includ'd files in a
source file.  This will be useful for adding #include tracking to my
bash support, and should let me implement a mini-libclang so I don't
have a hard libclang dependency.

pinclude is a simple program that exposes this functionality for
bash scripts and such.