Prevent pinclude from dying on non-existing headers
When pinclude (or libpinclude) ran into a header file that was
previously abandon the entire parse.

This old behavior is very different from what C compilers seem to do:
they instead silently skip the offending line.  This change fixes
pinclude to have the same behavior as GCC, and also adds a test case
to make sure the behavior remains consistant.