Support -L in pscalac
This is really just yet another copy of GCC's interface.  This
specifies a library search path where the scala compiler looks for
jars.

Note that, because of exactly how scala works, you have to specify
this to the complier as opposed to the linker.  This has to be managed
by the build system manually.