Automatically generate version number
The idea here is that I'll automatically tag this pconfigure version
with the output of 'git describe', as opposed to forcing me to
manually do everything myself.  I still have to manually update the
canonical version (the one used by package managers) because the git
stuff isn't included in the tarballs, but this should be good enough
for now.