Support "-S..." in pkg-config
The pkg-config support is still really a gaint sed script.  This loads
the sed from a file (dynamically) rather than baking it into the
Makefile.  The idea is that I can now build pkg-config files that have
dependencies that are satisfied by the Makefile, as opposed to by
pconfigure.

This is all part of my quest to only define the version number of
pconfigure in a single place and have everything else auto-generated.