Eleminate chisel l->deps
I actually have no idea what I was doing here: for some reason I was
storing the list of dependencies for a single chisel context inside
the chisel language, then iterating through that to find one of the
sources.

The fix is to simply use "c->full_path", which is what I should have
been using in the first place!