Add a simple pconfigure test
I'm not entirely sure this is the correct way to go about testing
pconfigure, but I guess it's OK for now.  Essentially this just
invokes the compiler to build things exactly as pconfigure would in
the real world.

I'm still thinking that some sort of approach using mock binaries (ie,
GCC and friends) could get us better coverage: in other words, it's
going to be very tricky to shake out edge cases this way.

The advantage is that it'll actually run the compiler, so we can test
things like options on different systems (which have been a problem).