Support dependency tests in pconfigure harness
This allows tests to contain a simple script, called "./update", that
is run between invocations two invocations of make and changes the
source in some manner.  The idea is that we use this script to test
pconfigure's dependency handling: build the code, make a change, and
then build again.  If the code update makes some easily detectable
change then it cat be diffed against the gold test output for
correctness.