Fix build on RedHat
I was really hoping to avoid this: essentially it forces
"_XOPEN_SOURCES 500" on everything that includes "lang/ftwfn.h".  The
problem is that this is going to cause problems because it'll end up
relying on the order things are included in.

Of course, since this doesn't reproduce on Gentoo, I'm never going to
find the bug before pushing.  I guess that's why I have Hudson :).