Add a sleep in test/pconfigure/harness.bash
I think the problem here is that my system's time resolution isn't
good enough and I end up missing a dependency.  This only happens on
palmer-desktop (which is really out of date, and also runs tmpfs), and
when I manually look at the generate makefile it seems correct.

Adding this one second sleep ensures that we wait for a system timer
tick (in reality, many of them) between the file modifications, which
makes this work.