Add bin/ to PATH during pconfigure tests
This is a bit of a hack: essentially the problem is that the
"pconfigure" tests actually depend on a bunch of the binaries are part
of this package.  Without this patch these aren't tested as part of
the pconfigure tests (and when pconfigure isn't installed, they're not
even found at all).

I think I really need to add a mechanism for doing this to pconfigure
itself, as this won't properly handle dependencies and such.  I guess
for now it's at least kind of OK.