Add sleeps in test harness
I believe the problem is that I'm running the touch command too close
to the previous make command, which is causing the modified file to
have the same mtime as the unmodified file.  This causes problems on
systems with looser filesystem resolution.

This has been by just sleeping a bit so the modifications are farther
apart.