Handle #ifdef, #ifndef, #else in libpinclude
This allows me to actually use pconfigure without clang.  I guess most
people have been using it for a long time this way, but now it's at
least useful for me because I can conditionally depend on C files
(which is really the crux of what pconfigure is all about).

Note that this doesn't handle #define -- it expects that all #define
commands are passes as "-D" on the command line.  This isn't the
greatest, but I think it'll have to do for now.  Properly handling
(I'd probably have to pull in talloc and stringlist from pconfigure)
and I don't currently have any code that uses it.

Also, note that this doesn't handle #if defined(...), as I'd have to
have a boolean logic solver for that.

There's a simple test case for this, but it's not great.  Building
pconfigure itself is probably the best test case, but it's not super
reliable (as the errors tend to cause extra extern/ libraries to get
pulled in).