Allow "SOURCES +=" under "HEADERS +="
This allows me to specify the following

  HEADERS += header.h
  SOURCES += header.h

which will copy (at install time only!) "src/header.h" into
"$PREFIX/include/header.h".  Some C++ projects really just want to
install all their internal headers (as it seems to be tricky to define
an export header for C++), this mechanism allows them to do that.

Note that this is technically backwards-incompatible, but I very much
doubt anyone was depending on the old behavior.