Allow SRCDIR anywhere
It doesn't make a whole lot of sense to just stick SRCDIR anywhere,
but there are at least a few places (ie, right under a BINARIES) that
it does make sense.

This is pretty liberal about allowing it places where you can break
your Configfile, but I guess that's life!