Add LIBEXEC command
This lets me add binaries to the libexec directory, which isn't in
path.  This also adds a C define for __PCONFIGURE__LIBEXEC, which is
the mechanism that I'm using to access those libraries.

The __PCONFIGURE__LIBEXEC #define isn't great because it means that I
can only discover these binaries when the binary has been installed.
Ideally I want to find a mechanism for overloading the extra link on
installation in order to make this work better...