Work around a test dependency problem
If you were trying to test a C library by building C test harnesses
and linking them then there was the possibility of over-linking
things: you'd end up linking once against the library and then again
against every object in that library.  This happens because of the
header/object resolution, so for now dependencies aren't managed at
all in tests.

This should probably be fixed -- I think the correct answer is
probably to restrict headers to inside the test directory (or maybe
mask out SRCDIR dependencies?).  For now that's not going to happen...