Code dump for "pconfigure++", a new pconfigure
So this is actually super tiny: it's just a bit of scaffolding that
will allow me to actually start implementing this.  That said, the
whole of pconfigure++ is going to be a pretty big deal: I'm
essentially just rewriting the whole thing in C++11 because I've
started to realize how crusty pconfigure actually is inside.

One of the explicit goals here is to make pconfigure++ be a whole lot
more flexible.  I want to retain support for C-based languages, but
also graduate to more interesting things like Scala, Chisel, and LaTeX
(with the goal there being to pretty much replace tek) as well as the
ability to generate bash-based Makefile generators (like autoconf
can), proper support for cross-compiling and libexec, and the ability
to merge together multiple seperate projects in one build.

Essentially it's just designed to fix all the problems with pconfigure
and be cleaner, too :).