Add string_utils, and a test case
One goal of pconfigure++ is to build up some proper utilities that I
can keep reusing.  I'm pretty much always doing stuff with strings, so
I figured the first would be to build up a set of string utilities.
As this is stuff that should be used outside pconfigure, I'm going to
try and actually test it...