Implement "{BINARIES,GENERATE,LIBRARIES,SOURCES,TESTS} +="
This just dumps a generic context onto the stack, which I'll need to
handle later via some sort of backend.  I think this is a decent
seperation: first handle parsing the commands and eliminate all the
state between different commands, and then generate the Makefile based
on a whole bunch of contexts that are independent.

This also results in context being a sub-type of opts_target, as it
can now have opts assigned to it.  In order to have things be a bit
cleaner, I've removed the member variables from opts_target and made
it abstract -- this means I have to duplicate add_{compile,link}opt()
twice, but I think it's OK as they're doing subtly different things
that allow me to have "context" be full of public variables.