Maintain a list of the contexts to be output
These are the top-level contexts: the ones that should be stuck in
"all".  What this means is that languages get to define a bit more
about exactly what dependencies mean than in the old pconfigure, where
this was mostly handled the same way for everyone.

This should mean that the whole "does this language need to compile"
flag can go away and be replaced by just cleanly subclassing stuff,
which should (hopefully) be somewhat nicer.