Find a relevant language for each output context
The language discovery semantics are somewhat different than they were
in the old pconfigure: right now only the source files that are
explicitly listed as children of the output context are checked for
compatibility.

While this allows me to have a single language handle each output
context, it means there's going to be some amount of pain in getting
languages to link to each other.  I think the best way to handle this
is to have each language support a list of languages that they know
how to link into output contexts.  This involves some pain
(particularly if I end up adding extension libraries), but with any
luck it'll be OK -- and certianly more flexible than the old way of
doing it.