Add language.cares_about_static
There's a bit of logic in pconfigure that tries to look at a library's
extension and determine if it's a shared or static library.  Many
languages don't have this concept, this patch allows those languages
to turn off this logic.

This is particularly important for the upcoming man language, which
doesn't have simple file extension rules that fit the pattern.