Avoid a double-free bug in pconfigure
I don't know why I was freeing "o" here, but that was causing
everything after this to break since "o" is the top of most of the
memory in pconfigure.

My current fix is to just not free it as it's being freed later
anyway.

This patch should fix some of

  https://github.com/palmer-dabbelt/pconfigure/issues/3