Add map_util::range*
I frequently want to use std::multimap to store many things that I
will then iterate through, but you can't directly use those with
range-based for loops.  This adds a little wrapper that acts as a shim
between a multimap (though it's templated, so it may work with other
things that have range_equals) and range-based for loops.