Add libmakefile::exclusive_target
This is a special sort of target that keeps around some global
bookkeeping to ensure that only one of this type of command (where
types are just a string that's passed in) is ever running at a given
time.

I'm using this to replace SBT's lockfiles with something that make
understands.