Add a minimal Chisel implementation
This is currently a bit of a mess: many things are hard-coded that
shouldn't be (like that we're generating behavorial Verilog as a
library, for example) but it's at least able to generate Verilog from
Scala sources, so that's a start.

I'm going to try and figure out how I want to test things next, so I
can try to integrate that into pconfigure as soon as possible.