Ensure that Makefile targets are only emitted once from Chisel
It's going to be pretty common for Chisel stuff to output the same
target many times, since people (or at least I) want to see the
intermediate results in a sane location.  This patch prevents Chisel
from outputing multiple definitions when that happens.

This will break if there are hash collisions or (and more likely) if
some argument isn't hashed that effects the output.  I'm actively
trying to minimize the arguments that get hashed so things don't have
to get run too often, so there's a very real risk of that.