Implement "pconfigure++ --verbose"
This is largely a debugging aid, both for me when writing pconfigure and
for users when trying to debug their Configfiles.  This support is a bit
hacky -- the actual "verbose" flag is stored with a context, but since
there's no way to pass it around everywhere it's instead set for an
entire Makefile.