Move the global language list to a function
I don't actually understand what's going on here: when compiling with
pconfigure I get the expected behavior, but when using bootstrap.sh I
end up with the _global variable being different for each function (but
uniqe across invocations of that function).  This only happens on my
Gentoo machine, Travis is fine.

I'm going to have to go read up on what static means in C++, but this
fixes the problem so I'm just committing it for now.