Stop using a global language list
This was a big change: now the language list is stored per context,
instead of per command_processor.  I think this may have introduced a
bug (which is why the LANGUAGES got re-ordered in pconfigure), but it
also fixes one (see the test case) so I'm going with it for now.