Don't pass C++ options to bash tests of C++ binaries
This is kind of funky, but essentially what happens here is that a test
inherits the compile/link options of its parent binary.  Since I usually
write tests in bash and my programs in C++, this causes a lot of
bailouts in pbashc.

The fix is to check to see if C++ tests are written in C++, and if
they're not to drop those options.  I think this is a reasonable
solution, but it could be fragile if I want to do something like test a
C library's C++ API compatibility.