Add a test for HEADERS without SOURCES
This was kind of awkward, but it's necessary to maintain
pconfigure-legacy compatibility.  In legacy pconfigure, you could add
HEADERS that didn't have sources.  This would just cause the headers to
be installed, and I believe there were no constraints on what could be
installed that way.

Additionally, you could do this without adding any LANGUAGES line that
cooresponds to this sort of header behavior.

I implemented this in pconfigure++ with another implicit language,
called "implicit_h", that just supports HEADERS that don't have SOURCES.